# src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Bancaria de Transferencias",
    "description": "API REST para operaciones bancarias de transferencias, diseñada para fines académicos de Ingeniería de Servicios Web.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Servidor local de desarrollo" }
  ],
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/api/transferencias": {
      "get": {
        "tags": ["Transferencias"],
        "summary": "Consulta histórico de transferencias del cliente",
        "description": "Permite consultar el historial de transferencias realizadas por el cliente, con filtros y paginación.",
        "operationId": "listarTransferencias",
        "parameters": [
          { "name": "fechaDesde", "in": "query", "required": false, "schema": { "type": "string", "format": "date", "example": "2026-01-01" } },
          { "name": "fechaHasta", "in": "query", "required": false, "schema": { "type": "string", "format": "date", "example": "2026-02-17" } },
          { "name": "estado", "in": "query", "required": false, "schema": { "type": "string", "example": "COMPLETADA" } },
          { "name": "pagina", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "name": "tamanio", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 } }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de transferencias",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransferenciaPaginada" } } }
          },
          "401": {
            "description": "No autorizado o token inválido",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "post": {
        "tags": ["Transferencias"],
        "summary": "Crea una transferencia bancaria entre cuentas",
        "description": "Permite crear una transferencia entre cuentas activas, validando saldo, límite diario y detectando operaciones sospechosas.",
        "operationId": "crearTransferencia",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransferenciaRequest" },
              "example": {
                "cuentaOrigen": "1001-2345-6789-0001",
                "cuentaDestino": "1001-9876-5432-1000",
                "monto": 5000,
                "moneda": "DOP",
                "concepto": "Pago factura",
                "idBeneficiario": "BEN-001"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transferencia creada exitosamente",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransferenciaResponse" } } }
          },
          "400": {
            "description": "Solicitud inválida o error de validación",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } }
          },
          "403": {
            "description": "Acceso denegado o validación de negocio fallida",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "422": {
            "description": "Operación no procesable (operación sospechosa)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/transferencias/{idTransaccion}": {
      "get": {
        "tags": ["Transferencias"],
        "summary": "Obtiene el detalle de una transferencia",
        "description": "Retorna el detalle completo de una transferencia específica.",
        "operationId": "obtenerDetalleTransferencia",
        "parameters": [
          { "name": "idTransaccion", "in": "path", "required": true, "schema": { "type": "string", "example": "TRX-20260217-0001" } }
        ],
        "responses": {
          "200": {
            "description": "Detalle de la transferencia",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransferenciaDetalle" } } }
          },
          "404": {
            "description": "Transferencia no encontrada",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "No autorizado o token inválido",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/transferencias/{idTransaccion}/comprobante": {
      "get": {
        "tags": ["Transferencias"],
        "summary": "Obtiene un comprobante JSON de la transferencia",
        "description": "Retorna un comprobante con código de autorización y estado de la transferencia.",
        "operationId": "obtenerComprobanteTransferencia",
        "parameters": [
          { "name": "idTransaccion", "in": "path", "required": true, "schema": { "type": "string", "example": "TRX-20260217-0001" } }
        ],
        "responses": {
          "200": {
            "description": "Comprobante de la transferencia",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ComprobanteResponse" },
                "example": {
                  "idTransaccion": "TRX-20260217-0001",
                  "codigoAutorizacion": "AUT-998877",
                  "fecha": "2026-02-17T16:30:00Z",
                  "estado": "COMPLETADA",
                  "mensaje": "Transferencia procesada correctamente"
                }
              }
            }
          },
          "404": {
            "description": "Transferencia no encontrada",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "No autorizado o token inválido",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "codigo": { "type": "string", "example": "ERROR_INTERNO" },
          "mensaje": { "type": "string", "example": "Ha ocurrido un error" },
          "detalles": { "type": "object", "nullable": true },
          "timestamp": { "type": "string", "format": "date-time" },
          "path": { "type": "string" },
          "requestId": { "type": "string" }
        }
      },
      "ValidationErrorResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ErrorResponse" },
          {
            "type": "object",
            "properties": {
              "codigo": { "type": "string", "example": "ERROR_VALIDACION" },
              "detalles": { "type": "object", "example": { "monto": ["El monto debe ser mayor que 0"] } }
            }
          }
        ]
      },
...
      "TransferenciaRequest": {
        "required": ["cuentaOrigen", "cuentaDestino", "monto", "moneda", "concepto"],
        "type": "object",
        "properties": {
          "cuentaOrigen": { "type": "string", "minLength": 10, "maxLength": 25, "example": "1001-2345-6789-0001" },
          "cuentaDestino": { "type": "string", "minLength": 10, "maxLength": 25, "example": "1001-9876-5432-1000" },
          "monto": { "type": "number", "format": "double", "minimum": 0.01, "example": 5000 },
          "moneda": { "type": "string", "enum": ["DOP", "USD", "EUR"], "example": "DOP" },
          "concepto": { "type": "string", "maxLength": 140, "example": "Pago factura" },
          "idBeneficiario": { "type": "string", "nullable": true, "example": "BEN-001" }
        }
      },
      "TransferenciaResponse": {
        "type": "object",
        "properties": {
          "idTransaccion": { "type": "string", "example": "TRX-20260217-0001" },
          "estado": { "type": "string", "enum": ["COMPLETADA", "EN_REVISION", "RECHAZADA"], "example": "COMPLETADA" },
          "fecha": { "type": "string", "format": "date-time" },
          "monto": { "type": "number", "format": "double", "example": 5000 },
          "moneda": { "type": "string", "enum": ["DOP", "USD", "EUR"], "example": "DOP" },
          "concepto": { "type": "string", "example": "Pago factura" },
          "cuentaOrigen": { "type": "string", "example": "1001-2345-6789-0001" },
          "cuentaDestino": { "type": "string", "example": "1001-9876-5432-1000" }
        }
      },
      "TransferenciaDetalle": {
        "allOf": [
          { "$ref": "#/components/schemas/TransferenciaResponse" },
          {
            "type": "object",
            "properties": {
              "idBeneficiario": { "type": "string", "nullable": true, "example": "BEN-001" },
              "mensaje": { "type": "string", "example": "Transferencia procesada correctamente" }
            }
          }
        ]
      },
      "TransferenciaPaginada": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/TransferenciaResponse" } },
          "paginacion": { "$ref": "#/components/schemas/MetadataPaginacion" }
        }
      },
      "ComprobanteResponse": {
        "type": "object",
        "properties": {
          "idTransaccion": { "type": "string", "example": "TRX-20260217-0001" },
          "codigoAutorizacion": { "type": "string", "example": "AUT-998877" },
          "fecha": { "type": "string", "format": "date-time" },
          "estado": { "type": "string", "example": "COMPLETADA" },
          "mensaje": { "type": "string", "example": "Transferencia procesada correctamente" }
        }
      },
...
      "MetadataPaginacion": {
        "type": "object",
        "properties": {
          "pagina": { "type": "integer", "example": 1 },
          "tamanio": { "type": "integer", "example": 10 },
          "totalRegistros": { "type": "integer", "example": 25 },
          "totalPaginas": { "type": "integer", "example": 3 }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token"
      }
    }
  }
}